# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "API Documentation for the E-Commerce platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/admin/categories": {
            "get": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Get all categories (Admin)",
                "description": "Retrieve a list of all categories.",
                "operationId": "3fc11a81efbb03b2004237b1ca11fdb9",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Create a new category (Admin)",
                "description": "Add a new category with an optional image.",
                "operationId": "b78202d020b5c79443c4e8e74bcf50e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/categories/{id}": {
            "put": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Update a category (Admin)",
                "description": "Update an existing category.",
                "operationId": "0618d8c185e1790c0729b548cc5ff4d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Delete a category (Admin)",
                "description": "Delete a category from the database.",
                "operationId": "445038f33e36fda29d31166b5ea6e5c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/orders": {
            "get": {
                "tags": [
                    "Admin - Orders"
                ],
                "summary": "Get all orders (Admin)",
                "description": "Retrieve a paginated list of all orders.",
                "operationId": "9ef0f005173b194048f3456b2b0b9f2f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/status": {
            "put": {
                "tags": [
                    "Admin - Orders"
                ],
                "summary": "Update order status (Admin)",
                "description": "Update the status of an existing order.",
                "operationId": "3ece12a34dece9dac463aa9dd0589998",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "New status (e.g. pending, confirmed, processing, shipped, delivered, cancelled)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}